+++ src/kernels/reduced_stats.hpp
@@ -21,7 +21,7 @@
 namespace kernel {
   using namespace ntt;
 
-  template <SimEngine::type S, MetricClass M, StatsID::type F, unsigned short I = 0>
+  template <SimEngine::type S, MetricClass M, StatsID::type F, uint8_t I = 0>
   class ReducedFields_kernel {
     static constexpr auto D = M::Dim;
 
@@ -401,7 +401,7 @@ namespace kernel {
   class ReducedParticleMoments_kernel {
     static constexpr auto D = M::Dim;
 
-    const unsigned short     c1, c2;
+    const uint8_t            c1, c2;
     const array_t<int*>      i1, i2, i3;
     const array_t<prtldx_t*> dx1, dx2, dx3;
     const array_t<real_t*>   ux1, ux2, ux3;
@@ -416,27 +416,25 @@ namespace kernel {
     const real_t contrib;
 
   public:
-    ReducedParticleMoments_kernel(const std::vector<unsigned short>& components,
-                                  const array_t<int*>&               i1,
-                                  const array_t<int*>&               i2,
-                                  const array_t<int*>&               i3,
-                                  const array_t<prtldx_t*>&          dx1,
-                                  const array_t<prtldx_t*>&          dx2,
-                                  const array_t<prtldx_t*>&          dx3,
-                                  const array_t<real_t*>&            ux1,
-                                  const array_t<real_t*>&            ux2,
-                                  const array_t<real_t*>&            ux3,
-                                  const array_t<real_t*>&            phi,
-                                  const array_t<real_t*>&            weight,
-                                  const array_t<short*>&             tag,
-                                  float                              mass,
-                                  float                              charge,
-                                  bool     use_weights,
-                                  const M& metric)
-      : c1 { not components.empty() ? components[0]
-                                    : static_cast<unsigned short>(0) }
-      , c2 { (components.size() == 2) ? components[1]
-                                      : static_cast<unsigned short>(0) }
+    ReducedParticleMoments_kernel(const std::vector<uint8_t>& components,
+                                  const array_t<int*>&        i1,
+                                  const array_t<int*>&        i2,
+                                  const array_t<int*>&        i3,
+                                  const array_t<prtldx_t*>&   dx1,
+                                  const array_t<prtldx_t*>&   dx2,
+                                  const array_t<prtldx_t*>&   dx3,
+                                  const array_t<real_t*>&     ux1,
+                                  const array_t<real_t*>&     ux2,
+                                  const array_t<real_t*>&     ux3,
+                                  const array_t<real_t*>&     phi,
+                                  const array_t<real_t*>&     weight,
+                                  const array_t<short*>&      tag,
+                                  float                       mass,
+                                  float                       charge,
+                                  bool                        use_weights,
+                                  const M&                    metric)
+      : c1 { not components.empty() ? components[0] : static_cast<uint8_t>(0) }
+      , c2 { (components.size() == 2) ? components[1] : static_cast<uint8_t>(0) }
       , i1 { i1 }
       , i2 { i2 }
       , i3 { i3 }
